+++ Modules/Libraries/MITLibrariesSearchResultDetailViewController.m
@@ -495,7 +495,7 @@ static NSString * const kAvailableCopiesForDisplayKey = @"kAvailableCopiesForDis
             MITLibrariesHoldingLibraryHeaderCell *holdingHeaderCell = [self.tableView dequeueReusableCellWithIdentifier:kHoldingLibraryHeaderCellIdentifier];
             holdingHeaderCell.libraryNameLabel.text = locationName;
             holdingHeaderCell.libraryHoursLabel.text = @"Put hours here";
-            holdingHeaderCell.availableCopiesLabel.text = [NSString stringWithFormat:@"%i of %i available", availableCopiesNumber, totalCopiesNumber];
+            holdingHeaderCell.availableCopiesLabel.text = [NSString stringWithFormat:@"%li of %li available", (long)availableCopiesNumber, (long)totalCopiesNumber];
             return holdingHeaderCell;
         }
